requirements.txt
google-adk
pydantic
pytest